[ADD] class and function documentations
